upgrade to compileSdk 34 in progress

+++ app/src/main/java/at/hgz/vocabletrainer/VocableListActivity.java
@@ -59,16 +59,16 @@ public class VocableListActivity extends ListActivity {
 	
 	@Override
 	public boolean onOptionsItemSelected(MenuItem item) {
-	    switch (item.getItemId()) {
-	        case R.id.addVocable:
-	            adapter.add(new Vocable(-1, -1, "", ""));
-	            setSelection(adapter.getCount() - 1);
-	            return true;
-	        case R.id.deleteDictionary:
-	            deleteDictionary();
-	            return true;
-	        default:
-	            return super.onOptionsItemSelected(item);
+		int id = item.getItemId();
+		if (id == R.id.addVocable) {
+			adapter.add(new Vocable(-1, -1, "", ""));
+			setSelection(adapter.getCount() - 1);
+			return true;
+		} else if (id == R.id.deleteDictionary) {
+			deleteDictionary();
+			return true;
+		} else {
+			return super.onOptionsItemSelected(item);
 	    }
 	}
